Used Roostayfish's Repo to add sickrage support
It works for me with kodi 14.1, sickrage, and couchpotato. As far as I
know the other stuff works, but I don't use it. I don't really know what
I'm doing. To be honest, I'm surprised it works for me. Let me know if
it doesn't work for you, but I probably can't fix it. Sorry.